requirements.txt
PyGithub>=2.1.1
python-dateutil>=2.8.2
cachetools>=5.3.2
rich>=13.7.0